Use chain id contants for li.fi inside project

+++ backend/db/model/user.go
@@ -18,7 +18,7 @@ type User struct {
 	WidgetUserColor    string    `json:"widget_user_color" gorm:"default:'#ffffff'"`
 	WidgetAmountColor  string    `json:"widget_amount_color" gorm:"default:'#22c55e'"`
 	WidgetMessageColor string    `json:"widget_message_color" gorm:"default:'#ffffff'"`
-	PreferredChainID   int       `json:"preferred_chain_id" gorm:"default:1"`
+	PreferredChainID   int64     `json:"preferred_chain_id" gorm:"default:1"`
 	PreferredAsset     string    `json:"preferred_asset_address" gorm:"default:'0x0000000000000000000000000000000000000000'"`
 	Description        string    `json:"description"`
 	BackgroundURL      string    `json:"background_url"`
